Move annotated title outside of data items
