+++ android/test/arch/src/test/kotlin/net/mullvad/mullvadvpn/test/arch/classes/DataClassTest.kt
@@ -1,8 +1,12 @@
 package net.mullvad.mullvadvpn.test.arch.classes
 
+import androidx.compose.runtime.Composable
 import com.lemonappdev.konsist.api.Konsist
+import com.lemonappdev.konsist.api.ext.list.functions
 import com.lemonappdev.konsist.api.ext.list.modifierprovider.withDataModifier
 import com.lemonappdev.konsist.api.ext.list.properties
+import com.lemonappdev.konsist.api.ext.list.withAllAnnotationsOf
+import com.lemonappdev.konsist.api.verify.assertEmpty
 import com.lemonappdev.konsist.api.verify.assertFalse
 import org.junit.jupiter.api.Test
 
@@ -23,4 +27,22 @@ class DataClassTest {
             // would fail because the nested class Mutable has a var property.
             .properties(includeNested = false)
             .assertFalse { it.isVar }
+
+    @Test
+    fun `ensure data classes do not have functions that are annotated with composable`() =
+        Konsist.scopeFromProject()
+            .classes(includeNested = true)
+            .withDataModifier()
+            .functions(includeNested = true)
+            .withAllAnnotationsOf(Composable::class)
+            .assertEmpty()
+
+    @Test
+    fun `ensure data classes do not have getters that are annotated with composable`() =
+        Konsist.scopeFromProject()
+            .classes(includeNested = true)
+            .withDataModifier()
+            .properties(includeNested = true)
+            .mapNotNull { if (it.getter?.text?.contains("@Composable") == true) it else null }
+            .assertEmpty()
 }
